Initial commit.
Installs an app that works as a service. It detects when a UHD device
is plugged into a USB port and loads the appropriate firmware and FPGA
images. It kills itself immediately and the service runs as a
background thread to do the loading. A notification is published
explaining the stages (loading firmware, loading fpga, done, or
error).

The FPGA and FW files are stored as assets in the APK and extracted to
/sdcard. This means we no longer have to push them ourselves or try to
carry around the correct versions against the UHD library version this
service links against.