Better config and setup.
Bundles in config.conf (GNU Radio config options), thrift.conf, and
volk_config (with basics for ARMv7 support).

Unpacks these and moves them to:
* /sdcard/.gnuradio/config.conf
* /sdcard/.gnuradio/thrift.conf
* /sdcard/.volk/volk_config

Also creates a /sdcard/uhd_usb_info.txt file that contains the fd and
usbfs_path discovered when loading the UHD firmware/FPGA files.

The fd is NOT the same fd that other apps, so they still need to
discover that themselves. This file is more meant for use as a marker
that the UHD device(s) are programmed and ready.